Changed UserService, fixed user account fields saving

+++ src/main/java/com/nikshcherbakov/vacanciesfinder/repositories/LocationRepository.java
@@ -4,5 +4,5 @@ import com.nikshcherbakov.vacanciesfinder.models.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
-    Location findLocationByLongitudeAndLatitude(float lon, float lat);
+    Location findLocationByLongitudeAndLatitude(double lon, double lat);
 }
